### 2023-10-30 #5990 @Kevin-Mattheus-Moerman
@AndresRTejedor thanks for this re-submission. Initially I rejected this submission based on its apparent small size from the software report over in https://github.com/openjournals/joss-reviews/issues/5958. There I perceived the work to be contained in <50 lines of Python, and <300 lines of C++ (the counts were however based on the `joss` paper branch). However, after your re-submission, with which you also additional information on the functionality/significance of this work, I have triggered this pre-review and have re-run the software report. This time I made sure the report is for the `main` branch. Now the numbers appear to satisfy our minimum requirements.

Furthermore, I've had a fellow member of the editorial board, @lucydot, check this submission in terms of scope, and she believes it may be in scope, and has offered to edit this work.

@AndresRTejedor, in relation to the above software report, what code portions would you say capture the core aspects of this work in terms of functionality and your original work? Is it mainly the Python and C++ code? Or are Fortran and JavaScript codes also your original developments? (I ask the latter as in some cases libraries from others are provided to support the project, and in some cases JavaScript is auto-generated to facilitate a GUI wrapper for instance)

For the moment, @AndresRTejedor, can you please address the above potentially missing DOIs? :point_up: Note also, that you can remove any entries from the `.bib` file that you do not end up using in the paper. Once you have updated the DOIs, you can rerun the reference check by calling: `@editorialbot generate pdf`.
