### 2023-10-30 #5990 @AndresRTejedor
Dear @lucydot

Thank you very much for reconsidering the resubmission of our paper.

Regarding Kevin comments:

-The code is entirely written by the authors of the work apart from the JavaScript which is an auto-generated code. More precisely:
The C++ code is the core of the software that must be included in LAMMPS to carry out the simulations either for the Mold Integration or the Lattice Mold techniques.
The Python code includes analysis to perform the thermodynamic integration (Mold Integration and Lattice Mold) as well as determining the average nucleation time (Lattice Mold).
The Fortran code includes several analysis and running options which are both very useful in the context of liquid-solid interfaces. In particular, for the purpose of this package, it performs the calculation of the largest cluster through a local order parameter to determine the crystallization of the system (Lattice Mold, but mostly Mold Integration).

-The DOIs are missing because the references are copied from Scholar that does not include DOIs. I will include all the DOIs as soon as possible thoughtout this morning. I will re-generate the pdf when everything is ready.

Many thanks for the possitive feeback. Looking forward to continuing the review process.

On behalf of all the authors.
Andres
