+++ joss/paper.bib
@@ -5,6 +5,7 @@
   volume={141},
   number={13},
   pages={134709},
+  doi = {10.1063/1.4896621},
   year={2014},
   publisher={AIP Publishing}
 }
@@ -14,6 +15,7 @@
   journal={The Journal of Chemical Physics},
   volume={158},
   number={19},
+  doi={10.1063/5.0139699},
   pages={194794},
   year={2023},
   publisher={AIP Publishing}
@@ -25,6 +27,7 @@
   volume={16},
   number={47},
   pages={25916--25927},
+  doi={10.1039/C4CP03948C},
   year={2014},
   publisher={Royal Society of Chemistry}
 }
@@ -34,6 +37,7 @@
   journal={The Journal of chemical physics},
   volume={149},
   number={7},
+  doi={10.1063/1.5018303},
   year={2018},
   publisher={AIP Publishing}
 }
@@ -44,6 +48,7 @@
   volume={13},
   number={44},
   pages={19807--19813},
+  doi={10.1039/C1CP22167A},
   year={2011},
   publisher={Royal Society of Chemistry}
 }
@@ -53,6 +58,7 @@
   journal={Biophysical Journal},
   year={2023},
   pages={2973-2987},
+  doi={10.1016/j.bpj.2023.03.006},
   publisher={Elsevier}
 }
 @article{kirkwood1949statistical,
@@ -62,6 +68,7 @@
   volume={17},
   number={3},
   pages={338--343},
+  doi={10.1063/1.1747248},
   year={1949},
   publisher={American Institute of Physics}
 }
@@ -72,6 +79,7 @@
   volume={11},
   number={1},
   pages={15241},
+  doi={10.1038/s41598-021-94309-y},
   year={2021},
   publisher={Nature Publishing Group UK London}
 }
@@ -81,6 +89,7 @@
   journal={Faraday discussions},
   volume={195},
   pages={569--582},
+  doi={10.1039/C6FD00141F},
   year={2016},
   publisher={Royal Society of Chemistry}
 }
@@ -90,6 +99,7 @@
   journal={SoftwareX},
   volume={1},
   pages={19--25},
+  doi={10.1016/j.softx.2015.06.001},
   year={2015},
   publisher={Elsevier}
 }
@@ -100,6 +110,7 @@
   volume={134},
   number={17},
   pages={174502},
+  doi={10.1063/1.3578182},
   year={2011},
   publisher={AIP Publishing}
 }
@@ -110,6 +121,7 @@
   volume={15},
   number={6},
   pages={246--253},
+  doi={10.1016/j.cossms.2011.07.003},
   year={2011},
   publisher={Elsevier}
 }
@@ -120,6 +132,7 @@
   volume={131},
   number={1},
   year={2009},
+  doi={10.1063/1.3153843},
   pages={014504},
   publisher={AIP Publishing}
 }
@@ -130,6 +143,7 @@
   volume={48},
   number={6},
   pages={1357--1368},
+  doi={10.1080/00268978300100971},
   year={1983},
   publisher={Taylor \& Francis}
 }
@@ -140,6 +154,7 @@
   volume={23},
   number={35},
   pages={19611--19626},
+  doi={10.1039/D1CP01784E},
   year={2021},
   publisher={Royal Society of Chemistry}
 }
@@ -150,6 +165,7 @@
   volume={130},
   number={11},
   pages={118001},
+  doi={10.1103/PhysRevLett.130.118001},
   year={2023},
   publisher={APS}
 }
@@ -160,6 +176,7 @@
   volume={142},
   number={19},
   pages={194709},
+  doi={10.1063/1.4921185},
   year={2015},
   publisher={AIP Publishing}
 }
[…]
