+++ joss/paper.md
@@ -55,7 +55,7 @@ bibliography: paper.bib
 
 # Summary
 
-The determination of the free energy cost of forming an interface between two phases, *i.e.* the interfacial free energy, is critical for characterizing a phase transition. In the particular case of liquid-to-solid transitions, obtaining this quantity is not straightforward; neither experimentally nor computationally. Here, we present a computational package integrated in the Molecular Dynamics open-source software LAMMPS (Large-Scale Atomic/Molecular Massively Parallel Simulator)[@Thompson2022], which enables the direct calculation of the interfacial free energy between arbitrarily complex crystal structures and liquids/solutions at coexistence conditions through the Mold Integration method [@espinosa2014mold]. Furthermore, the extension of this method for determining crystal nucleation rates---i.e., the probablity of a critical nucleus to emerge per unit of volume and time within a metastable phase---is also incorporated in the package, termed Lattice Mold method [@espinosa2016lattice]. Altogether, we detail here the required codes, scripts, and instructions for evaluating the two most challenging quantities determining the feasibility of a liquid-to-solid transition, implemented to be included in the MD package LAMMPS. 
+The determination of the free energy cost of forming an interface between two phases, *i.e.* the interfacial free energy, is critical for characterizing a phase transition. In the particular case of liquid-to-solid transitions, obtaining this quantity is not straightforward; neither experimentally nor computationally. Here, we present a computational package integrated in the Molecular Dynamics open-source software LAMMPS (Large-Scale Atomic/Molecular Massively Parallel Simulator)[@Thompson2022], which enables the direct calculation of the interfacial free energy between arbitrarily complex crystal structures and liquids/solutions at coexistence conditions through the Mold Integration method [@espinosa2014mold]. Furthermore, the extension of this method for determining crystal nucleation rates---i.e., the probablity of a critical nucleus to emerge per unit of volume and time within a metastable phase---is also incorporated in the package, termed Lattice Mold method [@espinosa2016lattice]. Altogether, we detail here the required codes, scripts, and instructions for evaluating the two most challenging quantities determining the feasibility of a liquid-to-solid transition, implemented to be included in the MD package LAMMPS.
 
 # Statement of need
 
