+++ joss/paper.bib
@@ -209,7 +209,7 @@
   volume={57},
   number={4},
   pages={941--971},
-  doi={},
+  doi={10.1016/j.actamat.2008.10.020},
   year={2009},
   publisher={Elsevier}
 }
@@ -263,7 +263,7 @@
   volume={117},
   number={24},
   pages={13238--13247},
-  doi={10.1073/pnas.1917569117 },
+  doi={10.1073/pnas.1917569117},
   year={2020},
   publisher={National Acad Sciences}
 }
@@ -329,6 +329,6 @@
   doi={10.1063/1.445633},
   publisher = {AIP},
   title = {Molecular dynamics investigation of the crystal--fluid interface. I. Bulk properties},
-  volume = {79}
+  volume = {79},
   year = {1983}
 }
