### 2023-11-06 #5990 @lucydot
Hi @AndresRTejedor -

@AndresRTejedor Thanks for your comments. I've re-read the paper:

- your statement of need section is long and contains scientific context. I suggest breaking this into two sections: `statement of need` (typically one or two paragraphs: what problem the software solves and target audience) and something along the lines of `scientific background`. Bear in mind you are towards the upper limit on paper length.
- Please fix the reference `DiPasquale2020shuttleworth?`

If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
