### 2023-11-06 2a9e644
New section in paper.md
