+++ joss/paper.md
@@ -57,9 +57,11 @@ bibliography: paper.bib
 
 The determination of the free energy cost of forming an interface between two phases, *i.e.* the interfacial free energy, is critical for characterizing a phase transition. In the particular case of liquid-to-solid transitions, obtaining this quantity is not straightforward; neither experimentally nor computationally. Here, we present a computational package integrated in the Molecular Dynamics open-source software LAMMPS (Large-Scale Atomic/Molecular Massively Parallel Simulator)[@Thompson2022], which enables the direct calculation of the interfacial free energy between arbitrarily complex crystal structures and liquids/solutions at coexistence conditions through the Mold Integration method [@espinosa2014mold]. Furthermore, the extension of this method for determining crystal nucleation rates---i.e., the probablity of a critical nucleus to emerge per unit of volume and time within a metastable phase---is also incorporated in the package, termed Lattice Mold method [@espinosa2016lattice]. Altogether, we detail here the required codes, scripts, and instructions for evaluating the two most challenging quantities determining the feasibility of a liquid-to-solid transition, implemented to be included in the MD package LAMMPS.
 
-# Statement of need
+# Scientific background
 
-Liquid-to-solid phase transitions are present in many physical and industrial processes such as freezing [@espinosa2016ice], nucleation [@mer1952nucleation], casting [@asta2009solidification] or welding [@asta2009solidification]. A thorough knowledge of the underlying thermodynamic processes is essential for understanding and controlling such phenomena. In particular, the interfacial free energy, defined as reversible work required to form a unit area of the interface between two coexisting phases, is crucial for determining the conditions that lead to a phase transition. The calculation of the interfacial free energy of liquid-vapour or liquid-liquid phases can be easily attained by analyzing the anisotropy of the stress tensor [@kirkwood1949statistical],[@walton1983pressure]. However, the same method leads to flawed results when used in systems involving solid phases [@DiPasquale2020shuttleworth]. 
+Liquid-to-solid phase transitions are present in many physical and industrial processes such as freezing [@espinosa2016ice], nucleation [@mer1952nucleation], casting [@asta2009solidification] or welding [@asta2009solidification]. A thorough knowledge of the underlying thermodynamic processes is essential for understanding and controlling such phenomena. In particular, the interfacial free energy, defined as reversible work required to form a unit area of the interface between two coexisting phases, is crucial for determining the conditions that lead to a phase transition. The calculation of the interfacial free energy of liquid-vapour or liquid-liquid phases can be easily attained by analyzing the anisotropy of the stress tensor [@kirkwood1949statistical],[@walton1983pressure]. However, the same method leads to flawed results when used in systems involving solid phases [@DiPasquale2020shuttleworth]. In this work, we present two different computational methods to help in attaining this problem. 
+
+# Statement of need
 
 The Mold Integration method was developed by Espinosa *et al.* [@espinosa2014mold] as a simple and effective approach to calculate the interfacial free energy of crystalline phases in coexistence with a liquid phase. This technique uses potential wells, that are arranged in the simulation box as a mold that replicates a given crystallographic plane of the solid of interest. The attractive wells are modeled through a square-well-like potential that is characterized by the well radius $r_w$, the well depth $\varepsilon_w$ and the well slope $\alpha$. The Mold Integration method consists of the following steps:
 
