+++ joss/paper.bib
@@ -332,7 +332,7 @@
   volume = {79},
   year = {1983}
 }
-@article{10.1063/5.0028219,
+@article{(DiPasquale2020shuttleworth,
     author = {Di Pasquale, Nicodemo and Davidchack, Ruslan L.},
     title = "{Shuttleworth equation: A molecular simulations perspective}",
     journal = {The Journal of Chemical Physics},
