### 2023-11-06 2fb369c
Update ne section paper.md
