+++ joss/paper.bib
@@ -332,7 +332,7 @@
   volume = {79},
   year = {1983}
 }
-@article{(DiPasquale2020shuttleworth,
+@article{DiPasquale2020shuttleworth,
     author = {Di Pasquale, Nicodemo and Davidchack, Ruslan L.},
     title = "{Shuttleworth equation: A molecular simulations perspective}",
     journal = {The Journal of Chemical Physics},
@@ -341,7 +341,6 @@
     pages = {154705},
     year = {2020},
     month = {10},
-    abstract = "{Even though the study of interfacial phenomena can be traced back to Laplace and was given a solid thermodynamic foundation by Gibbs, it appears that some concepts and relations among them are still causing some confusion and debates in the literature, particularly for interfaces involving solids. In particular, the definitions of the concepts of interfacial tension, free energy, and stress and the relationships between them sometimes lack clarity, and some authors even question their validity. So far, the debates about these relationships, in particular the Shuttleworth equation, have taken place within the framework of classical thermodynamics. In this work, we are offering to look at these concepts within the framework of statistical mechanics, which can be readily tested in Molecular Dynamics (MD) simulations. For a simple one component system of particles interacting via the Lennard-Jones potential, we calculate by the cleaving method the excess free energy of a solid–vacuum interface (solid surface) for systems in different states of tangential strain and compare the results to the calculation of surface stress via the difference of normal and tangential forces at the surface. As a result, we demonstrate consistency, within the statistical uncertainty, of the thermodynamic and statistical mechanical definitions of surface free energy and surface stress and how they are expressed via interaction-dependent quantities in MD simulations.}",
     issn = {0021-9606},
     doi = {10.1063/5.0028219},
     url = {https://doi.org/10.1063/5.0028219},
