### 2023-11-06 9c1177e
Update REFERENCE paper.md
