+++ joss/paper.md
@@ -60,7 +60,7 @@ The determination of the free energy cost of forming an interface between two ph
 # Statement of need
 
 Liquid-to-solid phase transitions are present in many physical and industrial processes such as freezing [@espinosa2016ice], nucleation [@mer1952nucleation], casting [@asta2009solidification] or welding [@asta2009solidification]. A thorough knowledge of the underlying thermodynamic processes is essential for understanding and controlling such phenomena. In particular, the interfacial free energy, defined as reversible work required to form a unit area of the interface between two coexisting phases, is crucial for determining the conditions that lead to a phase transition. The calculation of the interfacial free energy of liquid-vapour or liquid-liquid phases can be easily attained by analyzing the anisotropy of the stress tensor [@kirkwood1949statistical],[@walton1983pressure]. However, the same method leads to flawed results when used in systems involving solid phases [@DiPasquale2020shuttleworth]. For that reason, here we present two different computational methods that can be directly incorporated in the LAMMPS code to facilitate their functionality. 
-
+ 
 # Scientific background
 
 The Mold Integration method was developed by Espinosa *et al.* [@espinosa2014mold] as a simple and effective approach to calculate the interfacial free energy of crystalline phases in coexistence with a liquid phase. This technique uses potential wells, that are arranged in the simulation box as a mold that replicates a given crystallographic plane of the solid of interest. The attractive wells are modeled through a square-well-like potential that is characterized by the well radius $r_w$, the well depth $\varepsilon_w$ and the well slope $\alpha$. The Mold Integration method consists of the following steps:
