### 2023-11-06 #5990 @AndresRTejedor
Dear @lucydot ,

I have amended the reference `DiPasquale2020shuttleworth`. Also, I have divided the previous `statement of need` in two different sections to independently provide the `scientific background` of the methods we present in our package.

For the Referees I propose the following:

- lorenzo-rovigatti
- merrygoat

Many thanks
