### 2023-11-16 #5990 @lucydot
Thanks @AndresRTejedor for the suggestions and update.

@merrygoat, @mattwthompson, @marjanAlbouye - :wave: are you available to review this submission to JOSS?
