### 2023-11-22 #6083 @editorialbot
[REVIEW]: Mold: a LAMMPS package to compute interfacial free energies and nucleation rates
**Submitting author:** @AndresRTejedor (<a href="http://orcid.org/0000-0002-9437-6169">Andrés R. Tejedor</a>)
**Repository:** https://github.com/AndresRTejedor/Mold
**Branch with paper.md** (empty if default branch): joss
**Version:** LAMMPS 2023
