### 2023-11-27 #6083 @lucydot
I note that although editorial-bot has not found an open source license, this project is licensed GNU GPL (with LICENSE in the repo).
