### 2023-12-14 #6083 @AndresRTejedor
> @AndresRTejedor the `Version` listed in the first comment on this thread corresponds to the release version of your software (we recommend [semantic versioning](https://semver.org/)); can you let me know what this is?

Dear @lucydot, many thanks for your comment. We have checked the version of the software and it would correspond to the LAMMPS stable release of 2nd Aug 2023 (1.1.0) since we have just added a new functionality to this version.  We will change it in the repository accordingly.
