### 2023-12-17 AndresRTejedor/Mold#2 @mattwthompson
[FEATURE REQUEST] Merge `joss` branch into `main`?
There's a fair amount of divergence between the `main` and `joss` branches, which makes it unclear which state of the software the paper refers to.

**Describe the solution you'd like**
Consider merging `joss` into `main`, or, if it's important to keep the contents of the paper off of `main`, merge the opposite direction such that the `joss` branch always has the most recent changes from `main`.

Switching between branches is cumbersome and leads to some discrepancies, i.e. the `joss` branch not having a LICENSE file.

**Additional context**
