### 2023-12-20 #6083 @marjanAlbouye
Hi @AndresRTejedor, as part of my review, I am trying to  follow the instructions [here](https://andresrtejedor.github.io/Mold/example_MI_lj.html) to reproduce the results of the crystal fluid interfacial free energy. First, I would like to thank you and your team for providing detailed instructions  and documentations to showcase the functionalities of the Mold package.

As someone who has no experience with LAMMPS to run simulations, I was not able to proceed with these instructions (please see below). It would be great if you could add more instructions/information to the docs that could make the process easier for those (like me) who are not familiar with LAMMPS commands.

1- The bash script in `utils/MI/1.Optimal_r/Run.sh` requires  `LAMMPS.job` file to submit the jobs for different radius and velocity seeds. As stated in the docs, the `LAMMPS.job` file depends on the user's machine, however, there is no instruction about the details of this file and what commands should be included in it. It would be great if you could provide some example/template files.

2- Step 6 in the instructions, requires compiling a fortan file named `analysisNPT_clusterKoos_vec_rhodist_dellago_pressloc.10.f`. I tried to compile the file with `gfortran` command, however, since I was not able to run the simulations in the previous step due to the missing `LAMMPS.job` error, the fortan file did not compile.  Here is the first error:
```
analysisNPT_clusterKoos_vec_rhodist_dellago_pressloc.10.f:3497:9:

       end do
         1
Error: Expecting END SUBROUTINE statement at (1)
```

I'm hoping once I successfully run the simulations, this error would resolve. Also, it would make it easier for the users if you add the instructions for compiling the fortan file in the docs.

Thank you!
