### 2023-12-22 #6083 @AndresRTejedor
> > We have checked the version of the software and it would correspond to the LAMMPS stable release of 2nd Aug 2023 (1.1.0) since we have just added a new functionality to this version. We will change it in the repository accordingly.
>
> @AndresRTejedor - the version is the version of the package you are submitting (which I guess is compatible with, but distinct from, a LAMMPS release?)

Dear @lucydot, yes the version of our software would be version 1.0.0, and it is compatible with the last LAMMPS release.

Many thanks!
