+++ utils/MI/1.Optimal_r/order/analysisNPT_clusterKoos_vec_rhodist_dellago_pressloc.10.f
@@ -1,5 +1,177 @@
+c Version anterior 2jul03.f
+c Cambios:
+c   Se lee el parámetro alfa del modelo.inp
+c   Se incluye en parametros_NpT.inc la variable icutfour que
+c   controla el número de vectores de espacio recíproco que 
+c   tendrán en cuenta en las sumas de Ewald.
+c
+c
+c Versión anterior 28feb03.f
+c Cambios:
+c     Se añade un escaling nuevo que respeta la tetragonalidad de los
+c  lados a y b de la caja (iscale=3)
+c     Se cambia el formato de salida de alguna linea del fichero init.dat
+c
+c Versión anterior 6feb03.f
+c Cambios:    Se hace una g(r) promedio para tipo iguales de pares de 
+c          distancias (esto sólo valdrá para el agua SPC/E)
+c             Se posibilita el cálculo del factor de estructura para 
+c          distintos planos hkl (subrrutina order) y además descomentando
+c          una serie de líneas se elige el hkl que da mayor factor de
+c          estructura.
+c
+c Versión anterior: 17dicc.f
+c Cambio: Acumulador de energía Lennard-Jones
+c
+c Vesión anterior:11dicc.f
+c Cambio: quitar los parámetros de las rutinas de ewald
+c         que vengan heredados del programa de Bresme.
+
+c Version del 10-01-2002
+c
+c
+c Como correr un modelo duro ?
+c A) Descomentar las lineas comentadas que tienen la 
+c    palabra duro   c   duro 
+c B) Asignar un valor de sigma distinto de cero (hard diameter)
+c    pero asignar epsilon de todos los atomos igual a cero
+c C) Meter un cutoff pequeño en modelo.inp (si no hay cargas)
+c
+c Cambios que hay que hacer cuando se pasa de unidades reducidas a 
+c absolutas: 
+c A) Descomentar la parte en que las cargas se multiplican
+c por un factor de 408.7 (cargas que se leen en unidades de electrón).
+c B) Descomentar la línea en que la presión se multiplica por un factor
+c que la transforma de bares a kt/sigma**3.
+c C) Descomentar la línea en la que la densidad se pasa de gr/cm**3 a 
+c moléculas por unidad de volumen con cuidado de que la masa molecular 
+c que se utiliza sea la de la sustancia que se desea simular.
+c D) Tener en cuenta que lo mismo que se hace en la subrutina entrada
+c con la presión y la densidad hay que deshacerlo para que salga en las 
+c unidades deseadas. 
+c C) Si se va a hacer un batch es importante revisar que las unidades de las
+c variables que se meten en el fichero init.dat sean las mismas que se utilizan
+c en el simulación.inp. 
+c  
+c    Programa para mezclas pluricomponentes de moleculas lineares 
+c    flexibles
+c y que calcula unas gss ya normalizadas correctamente como en la
+c literatura ,da unos r12mu correctos
+c para el calculo de la presion, lo unico hay que dar el peso de 
+c cada interaccion. Solo calcula la interaccion del 1-2 no la del 2-1
+c
+c   Incluye intercambios de especie quimica entre dos moleculas
+c   
+c   1.) Incorpora cambios en order1
[…]
