+++ _sources/example_MI_lj.md.txt
@@ -92,11 +92,14 @@ path='../../'
 - `path`: path to `lj_mold.in` and `mold_100.lmp`. Absolute path is highly recommended.
 Also,  the bash file includes a submission command `sbatch LAMMPS.job`, but `LAMMPS.job` is not provided as it depends on the user machine. 
 
+````{note}
+The basic line to run the simulation is `./lmp_serial -i lj_moldint.in` where `lmp_serial` is the LAMMPS executable to run in serial. This line must be include in the submission file `LAMMPS.job` along with the appropiate commands that depend on the machine. This line also allows to run in the command line. Please  see the [LAMMPS documentation](https://docs.lammps.org/Run_basics.html) for further informarion in how to run LAMMPS.
+````
 
 6. The analysis for this step consists in determining if there is induction time, i.e. further energy is required for the formation of the interface (see {footcite:t}`espinosa2014mold`). 
 To do so, we recommend analyzing the resulting trajectory using the order parameter ${\bar{q}}_6$ ({footcite:t}`lechner2008accurate`) to determine the number of crystal-like particles in the slab. 
 The recommended value for such analysis is a threshold of ${\bar{q}}_6=0.34$ to distinguish solid-like from liquid-like particles (solid-like for ${\bar{q}}_6>0.34$, and liquid-like otherwise). The cut-off distance for such analysis is $1.35\sigma$. This distance is also used to identify molecules of solid that belong to the same solid cluster.
-We provide a Fortran program to apply the ${\bar{q}}_6$ order parameter to the resulting trajectories. You must compile the program `utils/MI/1.Optimal_r/order/analysisNPT_clusterKoos_vec_rhodist_dellago_pressloc.10.f` to get the executable `a.out`.
+We provide a Fortran program to apply the ${\bar{q}}_6$ order parameter to the resulting trajectories. You must compile the program `utils/MI/1.Optimal_r/order/analysisNPT_clusterKoos_vec_rhodist_dellago_pressloc.10.f` to get the executable `a.out`. You can use either `gfortran` or `ifort -r8` command to compile the fortran script.
 Then, we provide a bash file `utils/MI/1.Optimal_r/Analysis.sh` to run the analysis and get the evolution of the largest cluster throughout the simulation. This script must be run for each well radius in the directory above the simulations directory. 
 The program enters the directory for each seed (e.g. `8kT_0seed`), and the user must provide the well depth (variable `kT`) and the path to access all the files provided in `utils/MI/1.Optimal_r/order/` (variable `path`).
 As a result, one obtains different curves for the biggest cluster as a function of time for the different well radii in the file `nbig.data` that can be plotted to get the following figure:
