### 2024-01-12 #6083 @marjanAlbouye
> Thank you @marjanAlbouye for reporting the error. One of the authors has adapted the code so it can be compile with `gfortran` instead. Please try to compile again the fortran code and let me know if there is still any error. I have included a comment in the documentation in this regard.

@AndresRTejedor Thank you. I was able to compile the fortran code after I pulled the repository, thanks for the fix.
Following step 6 in [optimal radius calculation](https://andresrtejedor.github.io/Mold/example_MI_lj.html) documentation results in the following output, I'm not sure if it's an error or some expected behavior.

command:

`bash Analysis.sh`

I already changed the `path` variable in that file to `Mold/utils/MI/1.Optimal_r/order/`.

The output error is:

```
kT8_seed0
rm: cannot remove ‘todaslasconf.dat’: No such file or directory
rm: cannot remove ‘caja’: No such file or directory
rm: cannot remove ‘klkl’: No such file or directory
rm: cannot remove ‘kkk’: No such file or directory
At line 4539 of file analysisNPT_clusterKoos_vec_rhodist_dellago_pressloc.10.f (unit = 2, file = 'todaslasconf.dat')
Fortran runtime error: End of file

Error termination. Backtrace:
#0  0x7fbbf0b7dd36 in list_formatted_read_scalar
	at ../../../libgfortran/io/list_read.c:2309
#1  0x427c4b in ???
#2  0x41d80c in ???
#3  0x42d16c in ???
#4  0x42d1df in ???
#5  0x7fbbefe69554 in ???
#6  0x400f08 in ???
#7  0xffffffffffffffff in ???
...
```
Similar error appears for all the other seeds. However, inside each seed's directory there's `nbig_seed.data` file with two columns of numbers written in it. Does that mean the analysis run was successful and the results are written correctly?  Thank you.
