### 2024-01-15 #6083 @AndresRTejedor
@marjanAlbouye that's good news! From nbig_seed.data you can get the largest cluster as a function of time. If you plot the data you should get somthing similar to the figure provided in the documentation for each well radius.
