### 2024-01-16 #6083 @marjanAlbouye
@AndresRTejedor @lucydot  Here's my review/comments. Sorry about the delay!

## paper summary

Paper focuses on implementing mold integrator into LAMMPS package to directly calculate the interfacial free energy between crystal and liquid/solutions. Additionaly, it includes the Lattice method for determining crystal nucleation rate.

Authors did a good job explaining the necessity of calculating the interfacial free energy, providing scientific background for both the Mold integration method and the Lattice Mold technique. The paper also discusses the limitations of the commonly used GROMACS package with tabulated potentials to run simulations, reinforcing the case for a faster and simpler implementation, namely the MOLD package.

The paper also mentions two working examples included in the package documentation that demonstrate the mold integration and the lattice mold calculation functionalities to reproduce previously published results.

Overall, I found the paper easy to understand with strong statement of the need and background sections. Including working examples that reproduce previous results shows the authors’ commitment to building a reproducible package that could benefit other researchers in the community.

## Comments/Questions about the package:

- Seems like the MOLD package’s primary coding contributions is around implementing the two square-like potentials available in `Mold/src`. Are there any additional coding contributions made by the authors to LAMMPS for implementation of the mold integration and lattice mold technique? If yes, it would be great to briefly mention them in the paper as well.

- The package includes C++, Python and some Fortran code. While the functionalities of the package are tested to some extent in the two working examples, having unit tests is necessary to find bugs and assure code quality. It will also make the maintenance of the code easier. I highly suggest authors to implement some unit tests for the package.

- Regarding file organization in the package, are the script files in `Mold/utils` specifically implemented for the two examples (i.e. Mold integration for the 100 plane of the fcc crystal and Lattice mold calculation of Ice Ih nucleation rate)? If that’s the case, it might be better to move them to the `examples` folder to separate the main code from example codes and makes it less confusing.

- I was able to run the mold integration example with help from the author and I also noticed the documentation has improved with additional instructions. However, as mentioned earlier,  it would be beneficial to include more detailed instructions, particularly targeted at readers unfamiliar with LAMMPS. I don’t have a specific suggestion for improving the documentation right now, it’s just something to keep in mind for making the package more inclusive.

Thank you.
