### 2024-01-22 #6083 @lucydot
Thanks for your thorough review @marjanAlbouye; is there anything that is, in your view, blocking acceptance to JOSS? You mention testing. Unit tests are not absolutely necessary (but encouraged in most cases). Are there testing steps for objectively checking the expected functionality of the software (e.g., a sample input file to assert behaviour)? I assume yes given the tutorials you have worked through?
