### 2024-01-22 #6083 @mattwthompson
I got it started but was [quickly confused](https://github.com/AndresRTejedor/Mold/issues/2) with the significant divergence in the two branches - I don't mind the paper being in a separate branch but there's quite a bit of difference in the source code and I'm not sure which I'm supposed to be looking at
