### 2024-01-22 #6083 @AndresRTejedor
> @AndresRTejedor @lucydot Here's my review/comments. Sorry about the delay!
> ## paper summary
>
> Paper focuses on implementing mold integrator into LAMMPS package to directly calculate the interfacial free energy between crystal and liquid/solutions. Additionaly, it includes the Lattice method for determining crystal nucleation rate.
>
> Authors did a good job explaining the necessity of calculating the interfacial free energy, providing scientific background for both the Mold integration method and the Lattice Mold technique. The paper also discusses the limitations of the commonly used GROMACS package with tabulated potentials to run simulations, reinforcing the case for a faster and simpler implementation, namely the MOLD package.
>
> The paper also mentions two working examples included in the package documentation that demonstrate the mold integration and the lattice mold calculation functionalities to reproduce previously published results.
>
> Overall, I found the paper easy to understand with strong statement of the need and background sections. Including working examples that reproduce previous results shows the authors’ commitment to building a reproducible package that could benefit other researchers in the community.
> ## Comments/Questions about the package:
>
>     * Seems like the MOLD package’s primary coding contributions is around implementing the two square-like potentials available in `Mold/src`. Are there any additional coding contributions made by the authors to LAMMPS for implementation of the mold integration and lattice mold technique? If yes, it would be great to briefly mention them in the paper as well.
>
>     * The package includes C++, Python and some Fortran code. While the functionalities of the package are tested to some extent in the two working examples, having unit tests is necessary to find bugs and assure code quality. It will also make the maintenance of the code easier. I highly suggest authors to implement some unit tests for the package.
>
>     * Regarding file organization in the package, are the script files in `Mold/utils` specifically implemented for the two examples (i.e. Mold integration for the 100 plane of the fcc crystal and Lattice mold calculation of Ice Ih nucleation rate)? If that’s the case, it might be better to move them to the `examples` folder to separate the main code from example codes and makes it less confusing.
>
>     * I was able to run the mold integration example with help from the author and I also noticed the documentation has improved with additional instructions. However, as mentioned earlier,  it would be beneficial to include more detailed instructions, particularly targeted at readers unfamiliar with LAMMPS. I don’t have a specific suggestion for improving the documentation right now, it’s just something to keep in mind for making the package more inclusive.
>
>
> Thank you.

Dear @marjanAlbouye
Thank you for your positive feedback about the package. Please find below the reply and clarifications to your comments.
•	The MOLD package implements the square-well potential (in a continuous version as proposed by us (Espinosa et al., JCP 2014)) to the LAMMPS software as the fundamental tool to carry out both the mold integration and the lattice mold techniques. However, we also provide a detailed step-by-step manual, codes, analysis scripts, and input files to apply these methods and reproduce well-established values of the interfacial free energy and nucleation rates computed for two archetypal systems (water and Lennard-Jones) using our method. Hence, this package enables the implementation of a complex methodology which has been cited over the past 10 years more than 600 times in total, but that only a single group (which collaborate with us) has been able to implement with our help and by providing them with our programs. Since then, this group (Blas Lab) has published 3 articles in 3 years using this methodology ([Algaba, J. et al.]( https://doi.org/10.1016/j.jcis.2022.05.029), [Romero-Guzmán, C. et al.]( https://doi.org/10.1063/5.0139699), [Zerón, I. et al.]( https://doi.org/10.1016/j.actamat.2008.10.020)) and has become a world-leading group in surface free energy calculations of methane hydrates.  Hence, we believe that by making available all our codes, files, and instructions to implement the method, we open new venues for the the usage of these techniques to the scientific community.
•	We believe that the two different steps of the methods presented in this package implicitly serve as unit tests to check the functionalities of the code and the correct implementation/performance of the simulations. Nevertheless, following the reviewer advise, we will include more specific instructions in the documentation to specify which specific values along the two steps of the methods need to be reproduced to ensure the correct implementation and functioning of both Mold integration and the Lattice mold techniques. We thank the reviewer for this useful suggestion.
•	The script files included in ‘Mold/utils’ are indeed implemented specifically for the two examples. Following the reviewer advise, we have moved the codes to the examples folder and adapted the documentation accordingly. We thank the reviewer again for this useful suggestion.
•	Finally, we have also added some notes in the documentation to provide further details about running LAMMPS for researchers who might not be familiar with the LAMMPS Molecular Dynamics software. Moreover, we have detailed (see for instance the notes [here]( https://andresrtejedor.github.io/Mold/example_MI_lj.html#optimal-radius-calculation) and [here]( https://andresrtejedor.github.io/Mold/example_LM_mw.html#well-occupancy-curves-thermodynamic-integration)) more explicitly further instructions about LAMMPS set-up and running instead of referring to its documentation so we are more inclusive to a general potential user.
