### 2024-01-22 #6083 @AndresRTejedor
> I got it started but was [quickly confused](https://github.com/AndresRTejedor/Mold/issues/2) with the significant divergence in the two branches - I don't mind the paper being in a separate branch but there's quite a bit of difference in the source code and I'm not sure which I'm supposed to be looking at

Dear @mattwthompson

The `joss` must be independent to hold the paper itself according to the [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) of the journal. The code and documentation of the package is in the `main` branch.
