### 2024-02-05 #6083 @mattwthompson
Apologies for the delay - I have a few questions but the overall work serves a useful contribution to the field. I anticipate recommending acceptance but would first like to see the answers to some of my questions.

The authors describe a software package that implements two methods for estimating an interfacial free energy. They also implement a couple of pair potentials that seem to commonly be used in these studies. The paper is a brief and effective compliment to the software and helped me, somebody with background knowledge in the topic but no specific expertise, understand the context.

The software installs fine on my computer and runs simulations without immediate issue. I didn't run anything to convergence or deeply validate the details of the pair potentials.

The software is supplemented by documentation, hosted on GitHub. There are installations instructions which, however brief, were all I needed. There is another section on how exactly to call these new pair potentials from within LAMMPS, something that a user absolutely needs. Finally, there are two long, study-scale examples of using this software to implement free energy estimates.

Comments/questions

* There are some references to "square-well_-like_" potentials (emphasis mine). I don't know what the "-like" is doing here , as far as I can tell it is simply the potential I've always understood to be called "square well."
* The paper discusses a number of studies that use tabulated potentials with GROMACS. GROMACS has undergone some significant changes in the past ten years, including removing some features which have not yet been re-introduced.
  * The cited papers cover a large date range, and this paper cites the GROMACS 5 paper. Specifying which GROMACS version(s) were used would be useful; just the major version would be sufficient.
  * It is claimed that "... long-range corrections in energy and pressure cannot be applied across the calculation ..." As somebody who was frustrated trying to do something niche with tabulated potentials in GROMACS many years ago, I don't find this too surprising. Is this still the case? What version(s) does this apply to?
  * The authors claim that simulations with tabulated potentials in GROMACS are slower than simulations in LAMMPS with an analytical potential. This is a bit surprising to me, since GROMACS is typically quite a bit faster than LAMMPS and I understand the typical slowdown associated with tabulated potentials to be less. Could you expand a bit on the performance differences here? I'm mostly curious; I don't think the choice to implement this in LAMMPS was wrong.
* The LAMMPS-facing C++ code uses the term "Pocillo" a lot. I don't know what this is, and it prevented me from gleaning from a distance what it's meant to do. It seem to translate to "mug" which might be how a square-well pair potential is described here? Whether it's jargon I'm not familiar with or a Spanish term I'm also not familiar with, a short comment ("This class implements the pair potential of x y z ...") would help people understand the code. (I take no issue with non-English languages being mixed into source code, I just want to understand what the code means, no matter the language.) I apologize if I missed a comment or nested README file somewhere.
