+++ joss/paper.md
@@ -63,7 +63,7 @@ Liquid-to-solid phase transitions are present in many physical and industrial pr
  
 # Scientific background
 
-The Mold Integration method was developed by Espinosa *et al.* [@espinosa2014mold] as a simple and effective approach to calculate the interfacial free energy of crystalline phases in coexistence with a liquid phase. This technique uses potential wells, that are arranged in the simulation box as a mold that replicates a given crystallographic plane of the solid of interest. The attractive wells are modeled through a square-well-like potential that is characterized by the well radius $r_w$, the well depth $\varepsilon_w$ and the well slope $\alpha$. The Mold Integration method consists of the following steps:
+The Mold Integration method was developed by Espinosa *et al.* [@espinosa2014mold] as a simple and effective approach to calculate the interfacial free energy of crystalline phases in coexistence with a liquid phase. This technique uses potential wells, that are arranged in the simulation box as a mold that replicates a given crystallographic plane of the solid of interest. The attractive wells are modeled through a square-like well potential that is characterized by the well radius $r_w$, the well depth $\varepsilon_w$ and the well slope $\alpha$. The Mold Integration method consists of the following steps:
 
 1. Preparation of the configuration, by introducing the mold into the liquid phase.
 2. Simulation of the liquid-mold system at constant well depth and different well radii to determine the optimal radius that keeps the liquid particles inside the wells.
@@ -84,7 +84,7 @@ The nucleation rate is also a fundamental observable required to characterize nu
 
 In the bottom part of \autoref{fig:mold} we depict a schematic picture of the Lattice Mold technique, showing the different simulation steps to compute the nucleation rate of the system (see [@sanchez2022homogeneous] for further details). The nucleation free energy to form a sub-critical cluster is calculated from the well-occupancy vs well depth curve for each well radius. The average induction time is then calculated by running independent trajectories (at the highest well depth used in the previous step) and measure the time it takes for the pre-critical crystal nucleus to overcome the free energy barrier and percolate the whole system.
 
[…]
