### 2024-02-07 #6083 @AndresRTejedor
> Apologies for the delay - I have a few questions but the overall work serves a useful contribution to the field. I anticipate recommending acceptance but would first like to see the answers to some of my questions.
>
> The authors describe a software package that implements two methods for estimating an interfacial free energy. They also implement a couple of pair potentials that seem to commonly be used in these studies. The paper is a brief and effective compliment to the software and helped me, somebody with background knowledge in the topic but no specific expertise, understand the context.
>
> The software installs fine on my computer and runs simulations without immediate issue. I didn't run anything to convergence or deeply validate the details of the pair potentials.
>
> The software is supplemented by documentation, hosted on GitHub. There are installations instructions which, however brief, were all I needed. There is another section on how exactly to call these new pair potentials from within LAMMPS, something that a user absolutely needs. Finally, there are two long, study-scale examples of using this software to implement free energy estimates.
>
> Comments/questions
>
> * There are some references to "square-well_-like_" potentials (emphasis mine). I don't know what the "-like" is doing here , as far as I can tell it is simply the potential I've always understood to be called "square well."
> * The paper discusses a number of studies that use tabulated potentials with GROMACS. GROMACS has undergone some significant changes in the past ten years, including removing some features which have not yet been re-introduced.
>
>   * The cited papers cover a large date range, and this paper cites the GROMACS 5 paper. Specifying which GROMACS version(s) were used would be useful; just the major version would be sufficient.
>   * It is claimed that "... long-range corrections in energy and pressure cannot be applied across the calculation ..." As somebody who was frustrated trying to do something niche with tabulated potentials in GROMACS many years ago, I don't find this too surprising. Is this still the case? What version(s) does this apply to?
>   * The authors claim that simulations with tabulated potentials in GROMACS are slower than simulations in LAMMPS with an analytical potential. This is a bit surprising to me, since GROMACS is typically quite a bit faster than LAMMPS and I understand the typical slowdown associated with tabulated potentials to be less. Could you expand a bit on the performance differences here? I'm mostly curious; I don't think the choice to implement this in LAMMPS was wrong.
> * The LAMMPS-facing C++ code uses the term "Pocillo" a lot. I don't know what this is, and it prevented me from gleaning from a distance what it's meant to do. It seem to translate to "mug" which might be how a square-well pair potential is described here? Whether it's jargon I'm not familiar with or a Spanish term I'm also not familiar with, a short comment ("This class implements the pair potential of x y z ...") would help people understand the code. (I take no issue with non-English languages being mixed into source code, I just want to understand what the code means, no matter the language.) I apologize if I missed a comment or nested README file somewhere.

Dear @mattwthompson ,

Thank you for your positive feedback about the package. Please find below the reply and clarifications to all your comments and questions point by point.

- In some parts of the documentation, we use the term “square-like well potential” due to the difference with a proper square potential. While the latter is a discontinuous potential, and not suitable for MD simulations, our square-like well potential is continuous, and hence it can be differentiated to obtain the forces.  In this context, since our potential is continuous, and thus not exactly as a square-well one, we refer with the other term. You can find a further discussion in Fig.3 of [this paper] (https://doi.org/10.1063/1.4896621) which we cite in both the article and the documentation. We have added a line in the article explaining such nuance.

- Thank you for raising this relevant question. The version of GROMACS we used in those works was the v4.6.7. For later versions there are some incompatibility problems when using tabular potentials (at least we found them in GROMACS 5, 2016, and 2018). We have included a sentence stating the employed version of GROMACS that we used in our former published calculations using the Mold Integration and Lattice Mold techniques.
- To the best of our knowledge, for v4.6.7 no long-range corrections can be introduced when using tabular potentials. Regarding newer versions such as 2016, or 2018 (and probably also in more recent versions) the same issue still occurs. That is huge inconvenience since it disables direct comparison with standard simulations with long-range corrections using implicitly defined potentials.
- As the reviewer states, LAMMPS normally performs slightly slower simulations than GROMACS (i.e., for classical water models, LAMMPS is roughly 10% slower than GROMACS; but this also depends on the version of each package). However, when using a tabular potential in GROMACS vs. an implicit potential in LAMMPS it emerges a significant difference in speed up due to the following reason. The absence of energy-pressure long-range corrections requires to significantly enlarge the potential cut-offs (e.g., from 8.5Å to 14Å in water models or electrolyte solutions; please see [this](https://doi.org/10.1103/PhysRevLett.130.118001), and [this work](https://doi.org/10.1021/acs.jpcc.5b11221)  for obtaining consistent values of density, potential energy, or phase diagrams among simulations using regular cut-offs (i.e., 8.5 or 9Å) and long-range corrections vs. non-long-range corrections. Hence, defining the potential intrinsically in LAMMPS solves this problem enabling long-range corrections and the use of regular cut-offs that strongly improve the LAMMPS performance vs. that of GROMACS, becoming LAMMPS ∼1.8 times faster than GROMACS.

- Thank you for this comment. We used the term “pocillo” along the code with the meaning of “small well” in Spanish. We have clarified such meaning along the documentation as requested by the reviewer. Overall, we thank the reviewer for all the constructive comments to improve our manuscript/documentation.
