### 2024-02-07 #6083 @mattwthompson
@AndresRTejedor thanks for the responses

@lucydot save the lack of low-level tests, which aren't easy to write with LAMMPS plugins and from above I gather is not essential, I was able to check off every box in the checklist. I'm happy with the authors' responses to my quibbles and therefore recommend acceptance
