### 2024-02-14 #6083 @lucydot
## Adjustments to paper

Your reference lists are not formatted correctly. Where there are multiple references, use square brackets separated by semi-colons e.g. `[@anderson2020hoomd; @lammps_2022]`

In your paper I suggest re-phrasing for clarity:
- `In the particular case of liquid-to-solid transitions, obtaining this quantity is not straightforward neither experimentally nor computationally. ` --> `In the particular case of liquid-to-solid transitions, it is not straightforward to obtain this quantity through experiment or computation. `
- `Here, we present a computational package integrated in the Molecular Dynamics open-source software LAMMPS (Large-Scale Atomic/Molecular Massively Parallel Simulator)(Thompson et al., 2022), which enables the direct calculation of the interfacial free energy between arbitrarily complex crystal structures and liquids/solutions at coexistence conditions through the Mold Integration method (J. Espinosa et al., 2014).`  --> `Here, we present the computational package Mold, which is integrated in the Molecular Dynamics open-source software LAMMPS (Large-Scale Atomic/Molecular Massively Parallel Simulator)(Thompson et al., 2022). Mold enables direct calculation of the interfacial free energy between arbitrarily complex crystal structures and liquids/solutions at coexistence conditions through the Mold Integration method (J. Espinosa et al., 2014).`
- `implemented to be included in the MD package LAMMPS.` --> `implemented to work alongside the MD package LAMMPS.`
