+++ joss/paper.bib
@@ -1,6 +1,6 @@
 @article{espinosa2014mold,
   title={The mold integration method for the calculation of the crystal-fluid interfacial free energy from simulations},
-  author={Espinosa, JR and Vega, C and Sanz, E},
+  author={Espinosa, Jorge R and Vega, C and Sanz, E},
   journal={The Journal of chemical physics},
   volume={141},
   number={13},
