+++ joss/paper.md
@@ -27,10 +27,10 @@ authors:
     affiliation: 5
   - name: Nicodemo Di Pasquale
     orcid:  0000-0001-5676-8527
-    affiliation: 6
+    affiliation: "6,7"
   - name: Jorge Ramirez^[Corresponding author]
     orcid: 0000-0002-8946-3786
-    affiliation: 7
+    affiliation: 8
   - name: Jorge R. Espinosa^[Corresponding author]
     orcid: 0000-0001-9530-2658
     affiliation: 2
@@ -47,8 +47,10 @@ affiliations:
    index: 5
  - name: Department of Chemical Engineering, Brunel University London, United Kingdom
    index: 6
- - name: Department of Chemical Engineering, Universidad Politécnica de Madrid, Spain
+ - name: Dipartimento di Chimica Industriale, Università di Bologna, Italy
    index: 7
+ - name: Department of Chemical Engineering, Universidad Politécnica de Madrid, Spain
+   index: 8
 date: 01 Sep 2023
 bibliography: paper.bib
 ---
@@ -66,7 +68,7 @@ Liquid-to-solid phase transitions are present in many physical and industrial pr
 The Mold Integration method was developed by Espinosa *et al.* [@espinosa2014mold] as a simple and effective approach to calculate the interfacial free energy of crystalline phases in coexistence with a liquid phase. This technique uses potential wells, that are arranged in the simulation box as a mold that replicates a given crystallographic plane of the solid of interest. The attractive wells are modeled through a square-like well potential that is characterized by the well radius $r_w$, the well depth $\varepsilon_w$ and the well slope $\alpha$. The Mold Integration method consists of the following steps:
 
 1. Preparation of the configuration, by introducing the mold into the liquid phase.
-2. Simulation of the liquid-mold system at constant well depth and different well radii to determine the optimal radius that keeps the liquid particles inside the wells.
+2. Simulation of the liquid-mold system at constant well depth and different well radii to determine the limit of well radius (optimal radius) which supplies the exact free energy to form a crystal slab within the liquid.
 3. Calculation of the reversible work required to form a crystal slab at different well radii above the optimal well radius.
 4. Linear extrapolation of the interfacial free energy to the optimal well radius.
 
@@ -94,12 +96,12 @@ MOLD is a LAMMPS package for the calculation of interfacial free energy and nucl
 
 The code presented here can be found on [Github](https://github.com/AndresRTejedor/Mold), and it provides a detailed manual with the instructions to compile and use the package as well as with step-by-step examples to implement both the Mold Integration and the Lattice Mold techniques using the explicit square well potential. The package uses the `pair_style square/well` with three parameters that control the radius, the depth and the slope of the interaction potential. Additionally, we provide two detailed worked examples to reproduce already published results. 
 
[…]
