+++ joss/paper.bib
@@ -220,7 +220,7 @@
   volume={157},
   number={13},
   pages={134709},
-  doi={10.1016/j.actamat.2008.10.020},
+  doi={https://doi.org/10.1063/5.0101746},
   year={2022},
   publisher={AIP Publishing}
 }
