+++ joss/paper.md
@@ -86,9 +86,9 @@ The nucleation rate is also a fundamental observable required to characterize nu
 
 In the bottom part of \autoref{fig:mold} we depict a schematic picture of the Lattice Mold technique, showing the different simulation steps to compute the nucleation rate of the system (see [@sanchez2022homogeneous] for further details). The nucleation free energy to form a sub-critical cluster is calculated from the well-occupancy vs well depth curve for each well radius. The average induction time is then calculated by running independent trajectories (at the highest well depth used in the previous step) and measure the time it takes for the pre-critical crystal nucleus to overcome the free energy barrier and percolate the whole system.
 
-To date, the Mold Integration method has been widely used to calculate solid-liquid interfacial free energy of hard spheres [@espinosa2014mold], pseudo-hard spheres [@sanchez2021fcc], Lennard-Jones [@espinosa2014mold], SPC/E water with JC NaCl [@sanchez2023direct], NaCl (Tosi-Fumi) [@espinosa2015crystal], ice-water interfaces for TIP4P, TPI4P/2005 TIP4P/Ice and  mW models [@espinosa2016ice], and CO$_2$ hydrate water interfacial energy ([@zeron2022simulation], [@algaba2022simulation], [@romero2023effect]). The Lattice Mold method has been used to calculate the nucleation rate of pseudo-hard spheres and NaCl from its melt [@espinosa2016lattice], and mW and TIP4P/Ice water models [@sanchez2022homogeneous]. All these simulations have been carried out with GROMACS v4.6.7 [@abraham2015gromacs] and using a tabulated potential for the square-like attractive interaction between the wells and the fluid particles. Additionally, the typical run is more efficient in terms of computational resources when running on LAMMPS with the analytic form of the potential [@sanchez2022homogeneous], and the speed increases by a factor of $\sim1.8$. In fact, when using tabulated potentials in GROMACS, long-range corrections in energy and pressure cannot be applied across the calculation in GROMACS v4.6.7, and hence, larger cut-offs in the interaction potential need to be employed (e.g., from 8.5$\AA$ to 14$\AA$ in water models or electrolyte solutions). This is even more notable when studying complex systems such as atomistic models of water [@sanchez2022homogeneous] or hydrates [@zeron2022simulation]. Moreover, a tabulated potential needs to be generated for each run with different well parameters, which makes the process of simulation more complex and tedious. This package implements the square-like potential (please see Fig. 3 in  [@espinosa2014mold] for further discussion on the square-like well potential) in a explicit form on LAMMPS to speed up the simulation performance and simplify the implementation of both methods.
[…]
