### 2024-02-16 #6083 @AndresRTejedor
> list on the archive must match exactly that of the paper. C

Dear @lucydot

I have updated the repo to match the title and authors in the JOSS paper.
