+++ joss/paper.bib
@@ -220,7 +220,7 @@
   volume={157},
   number={13},
   pages={134709},
-  doi={https://doi.org/10.1063/5.0101746},
+  doi={10.1063/5.0101746},
   year={2022},
   publisher={AIP Publishing}
 }
@@ -343,6 +343,6 @@
     month = {10},
     issn = {0021-9606},
     doi = {10.1063/5.0028219},
-    url = {https://doi.org/10.1063/5.0028219},
+    url = {10.1063/5.0028219},
     eprint = {https://pubs.aip.org/aip/jcp/article-pdf/doi/10.1063/5.0028219/14698178/154705\_1\_online.pdf},
 }
