### 2024-02-28 #6083 @AndresRTejedor
@lucydot I have included Å directly into the markdown and it seems it worked. Thank you for noticing that.
