### 2024-03-04 #6083 @lucydot
Hi @AndresRTejedor  - there are double brackets `((` and `))` in the paper.md (occuring where there are lists of references). Please can you correct this?
