### 2024-03-04 #6083 @lucydot
It's also best to avoid the consecutive brackets on line 20/21 and the nested brackets on line 118/119. I'd suggest removing the brackets around the references by using square brackets around the sentence (see: https://stackoverflow.com/questions/64209134/r-markdown-suppress-parentheses-in-specific-citations), or re-phrasing.
