### 2024-03-06 #6083 @lucydot
@AndresRTejedor  -  hmm, not what I was expecting. I've tried fixing another way (removing square brackets from reference) and issued a PR. Could you merge and then we can check if that looks correct?

https://github.com/AndresRTejedor/Mold/pull/3
