### 2024-03-07 #6083 @lucydot
Hi @AndresRTejedor  - it hasn't.

I'll message the other editors to see how to include citations without the brackets. There is also the option of re-phrasing to avoid the citation within a bracket. We are very close to being finished!

FYI I'm going on leave for one week starting tomorrow, will check back here on the 18th. Please also note that if you make changes and want to verify correct, you can also check yourself using the `@editorialbot generate pdf` command.

Best,

Lucy
