### 2024-03-07 #6083 @AndresRTejedor
Hi @lucydot , thanks for your help! I hope it can be solved soon for publication!

Can I re-phrase those lines and finish it today?

Many thanks
