### 2024-03-19 #6083 @lucydot
@AndresRTejedor  - I have just spotted that the software version on the Zenodo archive (`Lammps`) does not match that which is set here (`v1.0.0`). To ensure consistency, you need to update the version number on Zenodo. To do this I think you can create a v1.0.0 release on Github and then Zenodo should issue a new doi.
