### 2024-03-21 #6083 @lucydot
@AndresRTejedor  you need to make the archive title and author list match exactly that of the JOSS submission. Please update. I'll re-paste the author to-do list from earlier message below so you can confirm all tasks complete:
