### 2024-03-21 #6083 @AndresRTejedor
Thanks @lucydot - I have checked the title and author list and now it matches those in the JOSS paper. Please let me know if you need anything else.
