+++ joss/paper.md
@@ -74,7 +74,7 @@ The Mold Integration method was developed by Espinosa *et al.* [@espinosa2014mol
 
 ![Top: Sketch of the thermodynamic pathway to compute the interfacial free energy using the Mold Integration method. Bottom: Schematic representation of the different steps in the Lattice Mold technique to compute the nucleation rate. In both schemes the empty black circles represent the potential wells that form the mold and the blue spheres represent the fluid particles. An extrapolation to a given well radius based on calculations performed at different radii is required in both methods.\label{fig:mold}](figs/joss.png){ width=95% }
 
-In the top part  of \autoref{fig:mold} we depict a sketch of the Mold Integration method with a generic solid plane and a liquid phase, showing the induction of the solid structure inside the liquid by the interactions of the liquid particles with the potential wells forming the mold (empty black circles). In stage 2, the depth of the wells is fixed to a value such that each well is filled with a fluid particle. Changes in the well radius leads to a modulation of the required free energy to create a crystal slab. A larger radius implies a longer induction time to form the interface so that the optimal radius is defined as the largest radius for which there is no induction time [@espinosa2014mold]. The thermodynamic integration (stage 3) is perfomed then for values above the optimal radius that allow reversibility (in the formation of the crystal slab) by varying the well depth for each radius. Extrapolating the interfacial energy to the optimal radius (stage 4) provides the estimation of the interfacial free energy. 
+In the top part  of \autoref{fig:mold} we depict a sketch of the Mold Integration method with a generic solid plane and a liquid phase, showing the induction of the solid structure inside the liquid by the interactions of the liquid particles with the potential wells forming the mold (empty black circles). In stage 2, the depth of the wells is fixed to a value such that each well is filled with a fluid particle. Changes in the well radius leads to a modulation of the required free energy to create a crystal slab. A larger radius implies a longer induction time to form the interface so that the optimal radius is defined as the largest radius for which there is no induction time [@espinosa2014mold]. The thermodynamic integration (stage 3) is performed then for values above the optimal radius that allow reversibility (in the formation of the crystal slab) by varying the well depth for each radius. Extrapolating the interfacial energy to the optimal radius (stage 4) provides the estimation of the interfacial free energy. 
 
 The nucleation rate is also a fundamental observable required to characterize nucleation phenomena. In this respect, the idea of using a mold to induce a crystal nucleus in the fluid leads to the Lattice Mold technique [@espinosa2016lattice]. In this method the nucleation rate is calculated by introducing a mold made of attractive wells whose positions are given by the crystallographic position of the solid at the same temperature for which the nucleation rate aims to be computed. In contrast with the Mold Integration method, the mold is not a crystal plane but a crystal cluster, typically with spherical shape, since this shape minimizes the interfacial area of the emerging cluster within the metastable fluid. In this case, the optimal radius of the well is reached in the limit beyond which more than a single fluid particle occupies a well. Hence, the thermodynamic integration is carried out for smaller radii than the optimal one. The mold must allow the system to nucleate (at maximum well depth) after a sufficiently long induction time, therefore ensuring reversibility across the thermodynamic integration step. This method can be summarized as follows:
 
