### 2024-03-25 #6083 @AndresRTejedor
@Kevin-Mattheus-Moerman thank you for your help. I have corrected the typo, but I don't find the `Creative Commons Attribution 4.0 International` line you mention. Where can I find it? What should I exactly put instead?

> @AndresRTejedor as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, however the below are some points that require your attention:
>
> * [ ]  Please manually edit the archive listed license to match your software license (it currently says `Creative Commons Attribution 4.0 International`).
> * [ v]  Check minor typo: `perfomed` -> `performed`
