### 2024-03-25 #6083 @Kevin-Mattheus-Moerman
@AndresRTejedor I have not done this myself in a while but I think if you manually edit the archive there is a field called Rights or license. Instead of the CC-BY license it should list your actual software license. Hope this helps.
