### 2024-03-26 #6083 @Kevin-Mattheus-Moerman
Yes this is for the ZENODO archive. I highlighted where the license is listed here in red:

![Screenshot from 2024-03-26 08-54-43](https://github.com/openjournals/joss-reviews/assets/8392709/5f5e7ae5-eb4e-4383-bc19-c7f7e5acf538)

Here is an animation (for some reason my cursor was not captured) of me editing one of my project entries, you go to edit then to license, then delete the CC-BY license and select your software license.

![Peek 2024-03-26 08-51](https://github.com/openjournals/joss-reviews/assets/8392709/dc6ced92-5a0f-4469-9a71-2b29d2e687b6)
