### 2024-03-26 #6083 @AndresRTejedor
@Kevin-Mattheus-Moerman many thanks for your help. I have already changed the license in the Zenodo release. Let me know if you need any further information.
